Special issue in MDPI Mathematics
I am a guest editor for a special issue in MDPI Mathematics:  Mathematical models in Evolutionary Ecology.  The idea is to publish  theoretical advances in  ecology, evolutionary biology, or  population genetics/genomics,  and  critical assessment of current theory through genomic data analysis.  The issue is now open for submissions,  and  we welcome contributions written for a broad audience.  In other words, if you want to contribute a theoretical paper,  you are kindly asked to try to make the theory  accessible  to evolutionary biologists or ecologists,  not just to the few mathematicians who might understand your calculation.